Self-audit round: 44 adversarially confirmed findings fixed
9-lens audit (shell, stratum, consensus Go, Python tooling, docker+CI,
docs-vs-reality, Lean FV, live ops, hygiene), every finding reproduced
by an independent verifier before fixing. Highlights:

- kawpow_core: DAG-offset modulus and fetch indices computed in 64-bit;
  the uint32 forms wrapped at a 16 GiB DAG (epoch 1921, ~5.9y in) and
  would have split the node from every stock GPU miner. Regression test
  pins cpp-kawpow semantics past the wrap point.
- ship_stratum_e2e: hermetic (port 3334, PID-scoped kills on both
  machines, chain-identity assert); the old version could kill the
  production sidecar and the WheelerGPU miner.
- join_wheeler_wsl: never wipes a datadir on a failed genesis probe
  (datadir lock read as "v1"); checkout converges on the pin every run.
- ci_prepare/ci_devnet: gates build in COREGETH_DIR=build/ci, never the
  tree the launchd production node execs from; interrupted fetches
  self-heal.
- dashboard: peer/miner strings escaped before innerHTML (stored XSS),
  orphaned burn_audit replays reaped, bind-before-spawn; launchd log
  rotation job; NAT-PMP failures now logged with a sentinel.
- stratum: authorize/submit race fixed, FIFO job eviction, sharediff=1
  target clamp; vet+race tests now a CI job.
- burn_audit models reverted txs and contract-touched accounts honestly.
- run-node.sh Art VIII guard keys on chain ID content, not filename;
  compose devnet is KawPow end to end (ethminer behind a profile).
- Docs: stratum image build steps, stock-client join warning (live-
  confirmed today), zero em dashes across all tracked markdown, LICENSE
  notes the vendored LGPL go-quai reference, GETH_INTEGRATION rewritten.

Gates: TestEverett/TestASERT/TestKawPow green in the isolated tree,
hermetic devnet e2e PASSED beside the live node, consistency gate 13/13,
live Wheeler audit wei-exact @3134 blocks, dashboard healthy post-reload.

+++ client/kawpow_engine.go
@@ -188,8 +188,8 @@ func kawpowFullFor(number uint64) []uint32 {
 }
 
 var (
-	kpDatasetMu    sync.Mutex // guards the mining DAG build, separate from
-	                          // kpMu so it never blocks verification
+	kpDatasetMu sync.Mutex // guards the mining DAG build, separate from
+	// kpMu so it never blocks verification
 	kpDataset      []uint32
 	kpDatasetEpoch uint64 = ^uint64(0)
 )
@@ -199,9 +199,9 @@ var (
 func kawpowComputeFull(hash []byte, nonce uint64, number uint64) ([]byte, []byte) {
 	e := kawpowEpochFor(number)
 	ds := kawpowFullFor(number)
-	lookup := func(index uint32) []byte {
+	lookup := func(index uint64) []byte {
 		mix := make([]byte, hashBytes)
-		for i := uint32(0); i < hashWords; i++ {
+		for i := uint64(0); i < hashWords; i++ {
 			binary.LittleEndian.PutUint32(mix[i*4:], ds[index+i])
 		}
 		return mix
